Fixed compile warning for 1Wire/Teleinfo Updated OZW to Dev branch

+++ hardware/1Wire/1WireByOWFS.cpp
@@ -182,7 +182,7 @@ void C1WireByOWFS::SetLightState(const std::string& sId,int unit,bool value, con
    case digital_potentiometer:
    {
 	   writeData(device, "chargepump", "1");
-	   unsigned int wiper = level * (255.0 / 15.0);
+	   unsigned int wiper = static_cast<unsigned int>(level * (255.0 / 15.0));
 	   writeData(device, "wiper", boost::to_string(wiper));
 	   break;
    }
